issue#70: customer app, wishlist model def
