--- .env
@@ -1,2 +1,0 @@
-hello
-hello
